adding some more support for prompts and openai
