+++ src/jurymind/core/prompts/__init__.py
@@ -1,0 +1,3 @@
+from jurymind.core.prompts.base import DEFAULT_SYSTEM_PROMPT, BasePromptTemplate
+
+__all__ = ["DEFAULT_SYSTEM_PROMPT", "BasePromptTempalate"]
