+++ src/jurymind/eval/base.py
@@ -1,0 +1,28 @@
+"""Base LLM evaluator class and modules"""
+
+
+class BaseEvaluator:
+    def evaluate():
+        pass
+
+
+class BleuScoreEvaluator(BaseEvaluator):
+    def evaluate():
+        pass
+
+
+class RougeScoreEvaluator(BaseEvaluator):
+    def evaluate():
+        pass
+
+
+class OpenAiEvaluator(BaseEvaluator):
+    pass
+
+
+class JudgeEvaluator(BaseEvaluator):
+    pass
+
+
+class JuryEvaluator(BaseEvaluator):
+    pass
