--- src/jurymind/eval/llm_evaluator.py
@@ -1,20 +1,0 @@
-"""Base LLM evaluator class and modules"""
-
-class BaseEvaluator:
-    def evaluate():
-        pass
-
-class BleuScoreEvaluator(BaseEvaluator):
-    pass
-
-class RougeScoreEvaluator(BaseEvaluator):
-    pass
-
-class OpenAiEvaluator(BaseEvaluator):
-    pass
-
-class JudgeEvaluator(BaseEvaluator):
-    pass
-
-class JuryEvaluator(BaseEvaluator):
-    pass
